+++ lib/Control/Monad/Dep/Has.hs
@@ -14,9 +14,9 @@
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ViewPatterns #-}
 
--- | This module provides a general-purpose \"Has\" class favoring a style in
--- which the components of the environment come wrapped in records or newtypes,
--- instead of being bare functions.
+-- | This module provides a general-purpose 'Has' class favoring a style in
+--   which the components of the environment come wrapped in records or newtypes,
+--   instead of being bare functions:
 --
 -- >>> :{
 --  type Logger :: (Type -> Type) -> Type
@@ -41,67 +41,79 @@
 --      repository :: Repository m,
 --      controller :: Controller m
 --    }
---  instance Has Logger m (Env m)
---  instance Has Repository m (Env m)
---  instance Has Controller m (Env m)
---  --
---  mkController :: MonadDep [Has Logger, Has Repository] d e m => Controller m
+--  -- instance Has Logger m (Env m)
+--  -- instance Has Repository m (Env m)
+--  -- instance Has Controller m (Env m)
+--  :}
+--  
+-- 'Has' can be used in combination with 'MonadDep', like this:
+--
+-- >>> :{
+--  mkController :: MonadDep [Has Logger, Has Repository] d env m => Controller m
 --  mkController =
---    Controller \url -> useCall \call -> do
+--    Controller \url -> 
+--      useEnv \(asCall -> call) -> do
+--        call log "I'm going to insert in the db!"
+--        call select "select * from ..."
+--        call insert [1, 2, 3, 4]
+--        return "view"
+-- :}
+--
+-- 'Has' can also be used independently of 'MonadReader' or 'MonadDep'. Here
+-- for example the environment is passed as a plain function argument, and @m@
+-- doesn't have any constraint other than 'Monad':
+--
+-- >>> :{
+--  mkController' :: (Monad m, Has Logger m env, Has Repository m env) => env -> Controller m
+--  mkController' (asCall -> call) =
+--    Controller \url -> do
 --      call log "I'm going to insert in the db!"
 --      call select "select * from ..."
 --      call insert [1, 2, 3, 4]
 --      return "view"
 -- :}
 --
--- The @adviseRecord@ and @deceiveRecord@ functions from the companion package
--- \"dep-t-advice\" can facilitate working with this style of components.
 --
 module Control.Monad.Dep.Has (
         -- * A general-purpose Has
         Has (..) 
-        -- * Component defaults
-    ,   Dep (..)
         -- * call helper
     ,   asCall
-    ,   useCall
+        -- * Component defaults
+    ,   Dep (..)
+--    ,   useCall
     ) where
 
 import Data.Kind
 import GHC.Records
 import GHC.TypeLits
 import Data.Coerce
-import Control.Monad.Reader
-import Control.Monad.Dep.Class
+-- import Control.Monad.Reader
+-- import Control.Monad.Dep.Class
 
 -- | A generic \"Has\" class. When partially applied to a parametrizable
 -- record-of-functions @r_@, produces a 2-place constraint that can be later
 -- used with "Control.Monad.Dep.Class".
 type Has :: ((Type -> Type) -> Type) -> (Type -> Type) -> Type -> Constraint
 class Has r_ m env | env -> m where
-  -- |  Given an environment @e@, produce a record-of-functions parameterized by the environment's effect monad @d@.
+  -- |  Given an environment @e@, produce a record-of-functions parameterized by the environment's effect monad @m@.
   --
   -- The hope is that using a selector function on the resulting record will
-  -- determine its type without the need for type annotations.
+  -- fix the record's type without the need for type annotations.
   --
-  -- (This will likely not play well with RecordDotSyntax. See also <https://chrisdone.com/posts/import-aliases-field-names/ this trick>.)
+  -- (This will likely not play well with RecordDotSyntax. See also <https://chrisdone.com/posts/import-aliases-field-names/ this import aliases trick for avoiding name collisions>.)
   dep :: env -> r_ m
   default dep :: (Dep r_, HasField (DefaultFieldName r_) env u, Coercible u (r_ m)) => env -> r_ m
   dep env = coerce . getField @(DefaultFieldName r_) $ env
 
+-- | Transforms an environment with suitable 'Has' instances into a \"helper\"
+--   function that looks in the environment for the arguments of other functions.
+--   Typically, the \"helped\" functions will be record field selectors.
+--
+--   Using 'asCall' in a view pattern avoids having to name the
+--   environment.
 asCall :: forall env m . env -> forall r_ x. Has r_ m env => (r_ m -> x) -> x
 asCall env = \f -> f (dep env)
-
--- | Avoids repeated calls to 'liftD' when all the effects in a function come
---   from the environment.
---
---   Similar to 'useEnv', but the callback receives a \"call\" function that
---   pre-applies any record selector with the correct record extracted from the
---   environment.
-useCall :: forall d env m y . (LiftDep d m, MonadReader env m) => ((forall r_ x . Has r_ d env => (r_ d -> x) -> x) -> d y) -> m y
-useCall usesCall = do
-  (asCall -> call) <- ask @env
-  liftD (usesCall call)
 
 -- | Parametrizable records-of-functions can be given an instance of this
 -- typeclass to specify the default field name 'Has' expects for the component